Spin off quaternion, vector, and orientation classes into their own files
