+++ src/python/adafruit_BNO055/jaiabot_imu.py
@@ -1,25 +1,20 @@
 #!/usr/bin/env python3
 from time import sleep
 from datetime import datetime
-import random
-import sys
 import argparse
 import socket
 import logging
 import time
 from math import *
 from dataclasses import dataclass
+from quaternion import Quaternion
+from vector3 import Vector3
+from orientation import Orientation
 
-parser = argparse.ArgumentParser(description='Read orientation, linear acceleration, and gravity from an AdaFruit BNO055 sensor, and publish them over UDP port')
-parser.add_argument('port', metavar='port', type=int, help='port to publish orientation data')
-parser.add_argument('-l', dest='logging_level', default='WARNING', type=str, help='Logging level (CRITICAL, ERROR, WARNING, INFO, DEBUG), default is WARNING')
-parser.add_argument('--simulator', action='store_true')
-parser.add_argument('--interactive', action='store_true')
-args = parser.parse_args()
 
 logging.basicConfig(format='%(asctime)s %(levelname)10s %(message)s')
 log = logging.getLogger('jaiabot_imu')
-log.setLevel(args.logging_level)
+
 
 try:
     import adafruit_bno055
@@ -30,103 +25,14 @@ except NotImplementedError:
     log.warning('NotImplementedError, so physical device not available')
 
 
-@dataclass
-class Orientation:
-    heading: float
-    pitch: float
-    roll: float
+parser = argparse.ArgumentParser(description='Read orientation, linear acceleration, and gravity from an AdaFruit BNO055 sensor, and publish them over UDP port')
+parser.add_argument('port', metavar='port', type=int, help='port to publish orientation data')
+parser.add_argument('-l', dest='logging_level', default='WARNING', type=str, help='Logging level (CRITICAL, ERROR, WARNING, INFO, DEBUG), default is WARNING')
+parser.add_argument('--simulator', action='store_true')
+parser.add_argument('--interactive', action='store_true')
+args = parser.parse_args()
 
-
-@dataclass
-class Vector3:
-    x: float
-    y: float
-    z: float
-
-
-    def dot(self, other):
-        return self.x * other.x + self.y * other.y + self.z * other.z
-
-
-    def cross(self, other):
-        return Vector3(
-            self.y * other.z - self.z * other.y,
-            -(self.x * other.z - self.z * other.x),
-            self.x * other.y - self.y * other.x
-        )
-
-
-    # k * v, scalar pre-multiplication
-    def __rmul__(self, other):
-        return Vector3(other * self.x, other * self.y, other * self.z)
-
-
-    # vector addition
-    def __add__(self, other):
-        return Vector3(self.x + other.x, self.y + other.y, self.z + other.z)
-
-
-@dataclass
-class Quaternion:
-    w: float
-    x: float
-    y: float
-    z: float
-
-
-    @staticmethod
-    def from_tuple(t: tuple):
-        return Quaternion(*t)
-
-
-    # Quaternion * Quaternion (multiplication)
-    def __mul__(self, other):
-        s1 = self.w
-        s2 = other.w
-        v1 = Vector3(self.x, self.y, self.z)
-        v2 = Vector3(other.x, other.y, other.z)
-        v = s1 * v2 + s2 * v1 + v1.cross(v2)
-
-        return Quaternion(
-            s1 * s2 - v1.dot(v2),
-            v.x, v.y, v.z
-        )
-
-
-    def to_euler_angles(self) -> Orientation:
-        DEG = pi / 180
-
-        # roll (x-axis rotation)
-        sinr_cosp = 2 * (self.w * self.x + self.y * self.z)
-        cosr_cosp = 1 - 2 * (self.x * self.x + self.y * self.y)
-        roll = atan2(sinr_cosp, cosr_cosp)
-
-        # pitch (y-axis rotation)
-        sinp = sqrt(1 + 2 * (self.w * self.y - self.x * self.z))
-        cosp = sqrt(1 - 2 * (self.w * self.y - self.x * self.z))
-        pitch = -2 * atan2(sinp, cosp) + pi / 2
-
-        # yaw (z-axis rotation)
-        siny_cosp = 2 * (self.w * self.z + self.x * self.y)
-        cosy_cosp = 1 - 2 * (self.y * self.y + self.z * self.z)
-        yaw = -atan2(siny_cosp, cosy_cosp)
-
-        if yaw < 0:
-            yaw += (2 * pi)
-
-        return Orientation(yaw / DEG, pitch / DEG, roll / DEG)
-
-
-    def unit_inverse(self):
-        # Assume unit quaternion!
-        return Quaternion(self.w, -self.x, -self.y, -self.z)
-
-
-    def apply(self, vector: Vector3):
-        v_quaternion = Quaternion(0, vector.x, vector.y, vector.z)
-        result_quaternion = self.unit_inverse() * v_quaternion * self
-        return Vector3(result_quaternion.x, result_quaternion.y, result_quaternion.z)
-
+log.setLevel(args.logging_level)
 
 class IMU:
