+++ src/python/adafruit_BNO055/orientation.py
@@ -1,0 +1,10 @@
+from dataclasses import dataclass
+
+
+
+@dataclass
+class Orientation:
+    heading: float
+    pitch: float
+    roll: float
+
