+++ src/python/adafruit_BNO055/vector3.py
@@ -1,0 +1,32 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Vector3:
+    x: float
+    y: float
+    z: float
+
+
+    def dot(self, other):
+        return self.x * other.x + self.y * other.y + self.z * other.z
+
+
+    def cross(self, other):
+        return Vector3(
+            self.y * other.z - self.z * other.y,
+            -(self.x * other.z - self.z * other.x),
+            self.x * other.y - self.y * other.x
+        )
+
+
+    # k * v, scalar pre-multiplication
+    def __rmul__(self, other):
+        return Vector3(other * self.x, other * self.y, other * self.z)
+
+
+    # vector addition
+    def __add__(self, other):
+        return Vector3(self.x + other.x, self.y + other.y, self.z + other.z)
+
+
